Rename integration test classes for Server
