+++ server/src/integration-test/java/com/scalar/db/server/DistributedStorageConditionalMutationIntegrationTestWithServer.java
@@ -1,12 +1,12 @@
 package com.scalar.db.server;
 
-import com.scalar.db.api.DistributedStorageMultiplePartitionKeyIntegrationTestBase;
+import com.scalar.db.api.DistributedStorageConditionalMutationIntegrationTestBase;
 import java.io.IOException;
 import java.util.Properties;
 import org.junit.jupiter.api.AfterAll;
 
-public class DistributedStorageServiceMultiplePartitionKeyIntegrationTest
-    extends DistributedStorageMultiplePartitionKeyIntegrationTestBase {
+public class DistributedStorageConditionalMutationIntegrationTestWithServer
+    extends DistributedStorageConditionalMutationIntegrationTestBase {
 
   private ScalarDbServer server;
 
@@ -31,5 +31,11 @@ public class DistributedStorageServiceMultiplePartitionKeyIntegrationTest
     if (server != null) {
       server.shutdown();
     }
+  }
+
+  @Override
+  protected int getThreadNum() {
+    // Since Deadlock error sometimes happens in MySQL, change the concurrency to 1
+    return 1;
   }
 }
