+++ server/src/integration-test/java/com/scalar/db/server/DistributedStorageMultiplePartitionKeyIntegrationTestWithServer.java
@@ -1,12 +1,12 @@
 package com.scalar.db.server;
 
-import com.scalar.db.api.DistributedStorageIntegrationTestBase;
+import com.scalar.db.api.DistributedStorageMultiplePartitionKeyIntegrationTestBase;
 import java.io.IOException;
 import java.util.Properties;
 import org.junit.jupiter.api.AfterAll;
 
-public class DistributedStorageServiceIntegrationTest
-    extends DistributedStorageIntegrationTestBase {
+public class DistributedStorageMultiplePartitionKeyIntegrationTestWithServer
+    extends DistributedStorageMultiplePartitionKeyIntegrationTestBase {
 
   private ScalarDbServer server;
