+++ server/src/integration-test/java/com/scalar/db/server/DistributedStorageSinglePartitionKeyIntegrationTestWithServer.java
@@ -1,12 +1,12 @@
 package com.scalar.db.server;
 
-import com.scalar.db.api.DistributedStorageConditionalMutationIntegrationTestBase;
+import com.scalar.db.api.DistributedStorageSinglePartitionKeyIntegrationTestBase;
 import java.io.IOException;
 import java.util.Properties;
 import org.junit.jupiter.api.AfterAll;
 
-public class DistributedStorageServiceConditionalMutationIntegrationTest
-    extends DistributedStorageConditionalMutationIntegrationTestBase {
+public class DistributedStorageSinglePartitionKeyIntegrationTestWithServer
+    extends DistributedStorageSinglePartitionKeyIntegrationTestBase {
 
   private ScalarDbServer server;
 
@@ -31,11 +31,5 @@ public class DistributedStorageServiceConditionalMutationIntegrationTest
     if (server != null) {
       server.shutdown();
     }
-  }
-
-  @Override
-  protected int getThreadNum() {
-    // Since Deadlock error sometimes happens in MySQL, change the concurrency to 1
-    return 1;
   }
 }
