+++ server/src/integration-test/java/com/scalar/db/server/SchemaLoaderIntegrationTestWithServer.java
@@ -1,6 +1,6 @@
 package com.scalar.db.server;
 
-import com.scalar.db.api.DistributedStorageAdminRepairTableIntegrationTestBase;
+import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
 import com.scalar.db.util.AdminTestUtils;
 import java.io.IOException;
 import java.util.Properties;
@@ -8,11 +8,10 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIf;
 
-public class DistributedStorageAdminServiceRepairTableIntegrationTest
-    extends DistributedStorageAdminRepairTableIntegrationTestBase {
+public class SchemaLoaderIntegrationTestWithServer extends SchemaLoaderIntegrationTestBase {
 
   private ScalarDbServer server;
-  private boolean isExternalServerUsed;
+  boolean isExternalServerUsed;
 
   @Override
   protected void initialize(String testName) throws IOException {
@@ -40,8 +39,8 @@ public class DistributedStorageAdminServiceRepairTableIntegrationTest
     return new ServerAdminTestUtils(properties);
   }
 
-  @Override
   @AfterAll
+  @Override
   public void afterAll() throws Exception {
     super.afterAll();
     if (server != null) {
@@ -53,24 +52,19 @@ public class DistributedStorageAdminServiceRepairTableIntegrationTest
   @Override
   @Test
   @DisabledIf("isExternalServerUsed")
-  public void repairTable_ForDeletedMetadataTable_ShouldRepairProperly() throws Exception {
-    super.repairTable_ForDeletedMetadataTable_ShouldRepairProperly();
+  public void createTableThenDropMetadataTableThenRepairTables_ShouldExecuteProperly()
+      throws Exception {
+    super.createTableThenDropMetadataTableThenRepairTables_ShouldExecuteProperly();
   }
 
   /** This test is disabled if {@link #isExternalServerUsed()} return true */
   @Override
   @Test
   @DisabledIf("isExternalServerUsed")
-  public void repairTable_ForTruncatedMetadataTable_ShouldRepairProperly() throws Exception {
-    super.repairTable_ForTruncatedMetadataTable_ShouldRepairProperly();
-  }
-
-  /** This test is disabled if {@link #isExternalServerUsed()} return true */
-  @Override
-  @Test
-  @DisabledIf("isExternalServerUsed")
-  public void repairTable_ForCorruptedMetadataTable_ShouldRepairProperly() throws Exception {
-    super.repairTable_ForCorruptedMetadataTable_ShouldRepairProperly();
+  public void
+      createTableThenDropMetadataTableThenRepairTablesWithCoordinator_ShouldExecuteProperly()
+          throws Exception {
+    super.createTableThenDropMetadataTableThenRepairTablesWithCoordinator_ShouldExecuteProperly();
   }
 
   @SuppressWarnings("unused")
